+++ deploy/config/microsoft/dockerconfig.yml
@@ -1,9 +1,5 @@
 kind: docker
 
-environmentName: foo
-
-installationPath: /tmp/tyger
-
 signingKeys:
   primary:
     public: ${HOME}/tyger_local_signing_key_public.pem
